after lesson

+++ src/test/java/task27/ArrayStackTest.java
@@ -3,7 +3,6 @@ package task27;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
 public class ArrayStackTest {
 
@@ -15,7 +14,7 @@ public class ArrayStackTest {
         as1.push(0);
         as1.push(1);
         as1.push(2);
-        assertEquals(2, as1.pop());
+        assertEquals(2, (int)as1.pop());
         assertEquals(false, as1.isEmpty());
 
     }
